move toggleBlockedSource to cmp data context
