+++ src/components/CampaignAnalytics/Placements.tsx
@@ -1,10 +1,8 @@
 import { Campaign, CampaignStatus, Placement } from 'adex-common'
 import CustomTable, { TableElement, TableRowAction } from 'components/common/CustomTable'
 import { getHumneSrcName } from 'helpers'
-import { useCallback, useMemo } from 'react'
+import { useMemo } from 'react'
 import { BaseAnalyticsData } from 'types'
-import { useAdExApi } from 'hooks/useAdexServices'
-import useCustomNotifications from 'hooks/useCustomNotifications'
 import InvisibilityIcon from 'resources/icons/Invisibility'
 import VisibilityIcon from 'resources/icons/Visibility'
 import { useCampaignsData } from 'hooks/useCampaignsData'
@@ -20,9 +18,7 @@ const Placements = ({
   placement: Placement
   campaign: Campaign
 }) => {
-  const { adexServicesRequest } = useAdExApi()
-  const { showNotification } = useCustomNotifications()
-  const { updateCampaignDataById } = useCampaignsData()
+  const { toggleBlockedSource } = useCampaignsData()
 
   if (!placements?.length) {
     return <div>No placement found</div>
@@ -81,48 +77,6 @@ const Placements = ({
     [placements, campaign.targetingInput.inputs.publishers.nin, placement, currencyName]
   )
 
-  const toggleBlocked = useCallback(
-    async ({ placementName, isBlocked, segment }: PlacementsTableElement['actionData']) => {
-      const blockedPublishers: Campaign['targetingInput']['inputs']['publishers'] = {
-        ...campaign.targetingInput.inputs.publishers,
-        apply: 'nin',
-        nin: isBlocked
-          ? [...campaign.targetingInput.inputs.publishers.nin].filter((x) => x !== segment)
-          : [...campaign.targetingInput.inputs.publishers.nin, segment]
-      }
-
-      const body: Pick<Campaign, 'pricingBounds' | 'targetingInput'> = {
-        pricingBounds: { ...campaign.pricingBounds },
-        targetingInput: {
-          ...campaign.targetingInput,
-          inputs: {
-            ...campaign.targetingInput.inputs,
-            publishers: blockedPublishers
-          }
-        }
-      }
-      try {
-        await adexServicesRequest('backend', {
-          route: `/dsp/campaigns/edit/${campaign.id}`,
-          method: 'PUT',
-          body,
-          headers: {
-            'Content-Type': 'application/json'
-          }
-        })
-        showNotification(
-          'info',
-          placementName,
-          `Successfully ${isBlocked ? 'unblocked' : 'blocked'}`
-        )
-        updateCampaignDataById(campaign.id)
-      } catch {
-        return showNotification('error', "Couldn't update the Campaign data!")
-      }
-    },
-    [adexServicesRequest, campaign, showNotification, updateCampaignDataById]
-  )
-
   const actions = useMemo(() => {
     const placementActions: TableRowAction[] = [
       CampaignStatus.active,
@@ -130,7 +84,8 @@ const Placements = ({
     ].includes(campaign.status)
       ? [
           {
-            action: (props: PlacementsTableElement['actionData']) => toggleBlocked(props),
+            action: (props: PlacementsTableElement['actionData']) =>
+              toggleBlockedSource(campaign.id, props.placementName, props.segment),
             label: ({ isBlocked, placementName }: PlacementsTableElement['actionData']) =>
               `${isBlocked ? 'Unblock' : 'Block'} "${placementName}"`,
             icon: ({ isBlocked }: PlacementsTableElement['actionData']) =>
@@ -140,7 +95,7 @@ const Placements = ({
       : []
 
     return placementActions
-  }, [campaign.status, toggleBlocked])
+  }, [campaign.id, campaign.status, toggleBlockedSource])
 
   return <CustomTable headings={headings} elements={elements} actions={actions} />
 }
